Remove separate OfferStatus component
